Slight modifications- nothing big
I wanted to do a more significant overall, with a more efficient
algorithm (like using pythagorean theorem), but decided not to. Instead
I just updated the for loop to do less tests (only check hexagonal if
pentagonal), and introduced triNum in the for loop initialization,
incrementing in the for loop end statement.