Solved problem 52!
After reading some of the online solutions I feel mine is not as
special, but I'm still proud of how I did this. I convert the current
long to a string, then to a character array, then sort that and convert
it back to a string. After doing that for each of the 1x 2x 3x .. 6x, I
compare them all (nested ifs).