YAHTZEE (real) beta release!
The last one was just a quick commit before I demo'ed it to Luke. This
is the real beta.

I've added a lot of features, and at this point I think it's safe to
say that the game is totally playable. The GUI updates accordingly,
it's user friendly, and you can play as many games as you like!

At this point I'm going to add an ideas list for future additions and
then start cranking through them, but I feel confident saying that I
started and finished Yahtzee in 24 hours.